fix: re expanded srNo field to show correct value
discarding the texbox widget but setting to progressbar again when collapsing the expander, that way on reexpand a new widget with new state is created for textbox so the initialValue is indeed the initial value. (which wasn't the case before, as the widget wasn't being destroyed on collapse. Added calls to notifyListeners() when ready field changes in the IncidentModel, previously, the model was silently  updating this even without notifylisteners(), as the previous lines had calls to mutate IncidentModel, that *did* do the notfiyListeners() call )